fix: smoke-test fallout — default query, schema columns, decimal handling
Three real bugs surfaced during the local end-to-end Submit smoke
test (simple-select ✅, total-revenue-per-customer wrong-order ✗,
right-order ✅):

1. SqlPlayground default query was hardcoded to "SELECT * FROM
   customers LIMIT 10;" for any problem with an initialSchema, but
   the users / hr schemas have no customers table. Replaced with a
   neutral comment that doesn't error on Run.

2. customers-by-country problem was authored with column `id`, but
   the ecommerce customers table uses `customer_id`. Updated the
   description, schemaDescription, and expectedOutput accordingly.
   The upsert's update block was also expanded so a re-seed
   refreshes content fields (otherwise the old broken row would
   persist).

3. lib/seed-data*.ts ECOMMERCE_SCHEMA / HR_SCHEMA used DECIMAL,
   which DuckDB-WASM Arrow returns to JS via toJSON() as the raw
   integer mantissa (1450.00 → 1450000 in JSON, breaking validation
   against expected 1450). Switched DECIMAL → DOUBLE in both
   schemas; semantically fine for a learning sandbox.

+++ lib/seed-data-hr.ts
@@ -13,7 +13,7 @@ CREATE TABLE employees (
 
 CREATE TABLE salaries (
     employee_id INTEGER,
-    amount DECIMAL(10,2)
+    amount DOUBLE
 );
 
 INSERT INTO departments VALUES (1, 'Engineering');
